myztrix/requirements.txt
addict==2.4.0
-e git+https://github.com/ArcadeAI/openai-agents-arcade.git@d5b553e6d7d7829f49b7890b32339421b011a64c#egg=agents_arcade
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
APScheduler==3.10.4
arcadepy==1.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
Automat==25.4.16
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
billiard==4.2.1
black==25.1.0
blinker==1.9.0
blis==1.3.0
boto3==1.38.27
botocore==1.38.27
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chromadb==1.0.12
chromedriver-autoinstaller==0.6.4
cleo==2.1.0
click==8.1.8
click-default-group==1.2.4
cloudpathlib==0.21.1
-e git+https://github.com/topoteretes/cognee.git@135e5b4651a1485567b59191c6dcdce39fa02bf1#egg=cognee
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.2
crashtest==0.4.1
cryptography==45.0.3
cssselect==1.3.0
curated-tokenizers==0.0.9
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
dateparser==1.2.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dlinfo==2.0.0
dlt==1.11.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
dulwich==0.22.8
durationpy==0.10
email_validator==2.2.0
emoji==2.14.1
espeakng-loader==0.2.4
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.11.0
fake-useragent==2.2.0
fastapi==0.115.7
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
Flask==2.3.3
flask-cors==6.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.2
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
google-api-core==2.25.0
google-api-python-client==2.118.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
graphistry==0.33.9
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.32.4
humanfriendly==10.0
humanize==4.12.3
hyperlink==21.0.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
installer==0.7.0
instructor==1.7.2
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==33.1.0
lancedb==0.21.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langfuse==2.60.8
langid==1.1.6
langsmith==0.3.45
language-tags==1.2.0
lazy_loader==0.4
limits==4.8.0
linkify-it-py==2.0.3
litellm==1.72.4
llama_cpp_python==0.3.9
llvmlite==0.44.0
lmstudio==1.3.1
lxml==5.4.0
mail-parser==4.1.3
makefun==1.16.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.9.4
mdit-py-plugins==0.4.2
mdurl==0.1.2
mergedeep==1.3.4
misaki==0.9.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.42.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
num2words==0.5.14
numpy==2.1.0
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.85.0
openai-agents==0.0.17
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
ordered-set==4.1.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
owlready2==0.47
packaging==24.2
palettable==3.3.3
pandas==2.3.0
parsel==1.10.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
pbs-installer==2025.6.6
pendulum==3.1.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
playsound3==3.2.4
pluggy==1.6.0
ply==3.11
plyer==2.1.0
poetry==2.1.3
poetry-core==2.1.3
posthog==4.6.1
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pure_eval==0.2.3
pwdlib==0.2.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.22.0
pync==2.0.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.6.0
pyperclip==1.9.0
PyPika==0.48.9
pypinyin==0.54.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-crontab==2.3.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==20.7
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyyaml_env_tag==1.1
queuelib==1.8.0
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.13.0
rfc3986==1.5.0
rich==13.9.4
rich-argparse==1.7.1
rpds-py==0.25.1
rsa==4.9.1
s3fs==2025.3.2
s3transfer==0.13.0
safetensors==0.5.3
schedule==1.2.1
scikit-learn==1.7.0
scipy==1.15.3
Scrapy==2.13.2
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.30.0
service-identity==24.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
soxr==0.5.0.post1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
sqlglot==26.28.0
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlite-vec==0.1.6
squarify==0.4.4
sse-starlette==2.3.6
starlette==0.45.3
streamlit==1.45.1
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
textual==3.3.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trove-classifiers==2025.5.9.12
Twisted==25.5.0
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
vaderSentiment==3.3.2
vibekit==0.1.0
vine==5.1.0
virtualenv==20.31.2
w3lib==2.3.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
win-precise-time==1.4.2
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
